requirements.txt
numpy
pillow
imageio[ffmpeg]
